rename to SegmentStarTable.
