+++ iris-common/src/main/java/cfa/vo/iris/sed/stil/SegmentStarTableAdapter.java
@@ -7,7 +7,7 @@ public class SegmentStarTableAdapter implements StarTableAdapter<ISegment> {
 
     @Override
     public StarTable convertStarTable(ISegment data) {
-        return new SegmentStarTableWrapper(data);
+        return new SegmentStarTable(data);
     }
 
 }
